+++ libcloudvagrant/common/__init__.py
@@ -18,19 +18,4 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
-import jinja2
-
-
-__all__ = [
-    "render",
-]
-
-
-loader = jinja2.PackageLoader("libcloudvagrant", "templates")
-env = jinja2.Environment(loader=loader, autoescape=False)
-
-
-def render(template_name, context, fname):
-    t = env.get_template(template_name)
-    with open(fname, "wt") as f:
-        f.write(t.render(context).encode("utf8"))
+"""Common definitions for all components."""