+++ libcloudvagrant/compute/driver.py
@@ -42,9 +42,10 @@ from libcloud.common.types import LibcloudError
 from libcloud.compute import base
 from libcloud.compute.types import DeploymentError, NodeState
 
-from libcloudvagrant import virtualbox
-from libcloudvagrant.catalogue import VagrantCatalogue
-from libcloudvagrant.types import (
+from libcloudvagrant.common import virtualbox
+from libcloudvagrant.common.catalogue import VagrantCatalogue
+from libcloudvagrant.common.types import VAGRANT
+from libcloudvagrant.compute.types import (
     VagrantImage,
     VagrantNetwork,
     VagrantNode,
@@ -58,8 +59,6 @@ __all__ = [
     "VagrantDriver",
 ]
 
-
-VAGRANT = "vagrant"
 
 # ``NODE_ONLINE_WAIT_TIMEOUT`` and ``SSH_CONNECT_TIMEOUT`` are needed in our
 # reimplementation of ``deploy_node``,